backend/requirements.txt
Django
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
docopt
drf-yasg
PyJWT